refactor: move some other files for not conflict with open_webui installation
