+++ .gitignore
@@ -1,2 +1,5 @@
 __pycache__
 .env
+
+.python-version
+dist/
