+++ dev.sh
@@ -1,2 +1,2 @@
 PORT="${PORT:-9099}"
-uvicorn main:app --port $PORT --host 0.0.0.0 --forwarded-allow-ips '*' --reload
+uvicorn open_webui.pipeline:app --port $PORT --host 0.0.0.0 --forwarded-allow-ips '*' --reload
