+++ open_webui/pipeline/__init__.py
@@ -10,3 +10,5 @@ try:
     load_dotenv(find_dotenv("./.env"))
 except ImportError:
     print("dotenv not installed, skipping...")
+
+from .main import app  # noqa
