+++ requirements-dev.lock
@@ -1,0 +1,100 @@
+# generated by rye
+# use `rye lock` or `rye sync` to update this lockfile
+#
+# last locked with the following flags:
+#   pre: false
+#   features: []
+#   all-features: false
+#   with-sources: false
+
+-e file:.
+annotated-types==0.7.0
+    # via pydantic
+anyio==4.3.0
+    # via httpx
+    # via starlette
+    # via watchfiles
+certifi==2024.2.2
+    # via httpcore
+    # via httpx
+    # via requests
+charset-normalizer==3.3.2
+    # via requests
+click==8.1.7
+    # via typer
+    # via uvicorn
+dnspython==2.6.1
+    # via email-validator
+email-validator==2.1.1
+    # via fastapi
+fastapi==0.111.0
+    # via open-webui-pipeline
+fastapi-cli==0.0.4
+    # via fastapi
+h11==0.14.0
+    # via httpcore
+    # via uvicorn
+httpcore==1.0.5
+    # via httpx
+httptools==0.6.1
+    # via uvicorn
+httpx==0.27.0
+    # via fastapi
+idna==3.7
+    # via anyio
+    # via email-validator
+    # via httpx
+    # via requests
+jinja2==3.1.4
+    # via fastapi
+markdown-it-py==3.0.0
+    # via rich
+markupsafe==2.1.5
+    # via jinja2
+mdurl==0.1.2
+    # via markdown-it-py
+orjson==3.10.3
+    # via fastapi
+pydantic==2.7.1
+    # via fastapi
+pydantic-core==2.18.2
+    # via pydantic
+pygments==2.18.0
+    # via rich
+python-dotenv==1.0.1
+    # via uvicorn
+python-multipart==0.0.9
+    # via fastapi
+pyyaml==6.0.1
+    # via uvicorn
+requests==2.32.2
+    # via open-webui-pipeline
+rich==13.7.1
+    # via typer
+shellingham==1.5.4
+    # via typer
+sniffio==1.3.1
+    # via anyio
+    # via httpx
+starlette==0.37.2
+    # via fastapi
+typer==0.12.3
+    # via fastapi-cli
+    # via open-webui-pipeline
+typing-extensions==4.12.0
+    # via fastapi
+    # via pydantic
+    # via pydantic-core
+    # via typer
+ujson==5.10.0
+    # via fastapi
+urllib3==2.2.1
+    # via requests
+uvicorn==0.29.0
+    # via fastapi
+uvloop==0.19.0
+    # via uvicorn
+watchfiles==0.21.0
+    # via uvicorn
+websockets==12.0
+    # via uvicorn
