+++ start.sh
@@ -2,4 +2,4 @@
 PORT="${PORT:-9099}"
 HOST="${HOST:-0.0.0.0}"
 
-uvicorn main:app --host "$HOST" --port "$PORT" --forwarded-allow-ips '*'
+uvicorn open_webui.pipeline:app --host "$HOST" --port "$PORT" --forwarded-allow-ips '*'
